requirements.txt
appdirs==1.4.4
asgiref==3.2.7
boto3==1.14.16
botocore==1.17.16
certifi==2020.4.5.1
click==7.1.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.2
django-environ==0.4.5
django-filter==2.3.0
django-forgiving-collectstatic==1.0.0
django-heroku==0.3.1
django-phone-field==1.8.1
django-storages==1.9.1
djangorestframework==3.11.1
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
Pillow==7.1.2
pipenv==2018.11.26
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
Werkzeug==1.0.1
whitenoise==5.0.1